qdl: Widen start_sector to u64 to avoid truncation on large storage
firehose_read_storage() and firehose_checksum_storage() took
start_sector as u32. GPT LBAs are u64; with 512-byte sectors any
partition starting beyond 2TiB was silently redirected to a wrong
address due to the truncating cast in the callers.

Widen the parameter and remove the as-u32 cast at the call site.

The programmer treats this value as u64, make it match.

Fixes: 7538a8da78a4 ("Initial commit")

+++ cli/src/programfile.rs
@@ -31,7 +31,7 @@ fn parse_read_cmd<T: QdlChan>(
         .unwrap()
         .parse::<u8>()
         .unwrap();
-    let start_sector = attrs.get("start_sector").unwrap().parse::<u32>().unwrap();
+    let start_sector = attrs.get("start_sector").unwrap().parse::<u64>().unwrap();
 
     if checksum_only {
         return firehose_checksum_storage(channel, num_sectors, phys_part_idx, start_sector);
